Make our Makefile generate fake SOs for other platforms
Currently, panda will complain if any file in META.info's resources
list is not present.  However, I need to be exhaustive in the list of
resources I provide (as far as the helper library goes), so I need to
list all of the possible shared object files.  This change makes it so
the shared object files not used on the current OS are just installed
as empty files