Fixes issues with lib name generation
This fixes a problem where automatic library name generation is
used, but part of the generated library name is hardcoded else-
where. This resulted in the build failing on windows, because
the hardcoded portion was not applicapable.

Note that while this allows Linenoise to be installed on windows
once again, it still does not work properly. This is because libs
that get installed have their file names mangled, but I think the
linenoise.dll file needs to have the same name? (some missing
symbol error gets thrown). OpenSSL had this problem, which it works
around with lib/OpenSSL/NativeLib.pm6